Minor updates to be compliant with libraries numpy-1.24 and scipy-1.10

+++ aveScan.py
@@ -80,7 +80,7 @@ def argument():
 
 args = argument()
 
-import scipy.io.netcdf as NC
+import scipy.io as NC
 import netCDF4 as NC4
 from commons.dataextractor import DataExtractor
 import glob
@@ -135,9 +135,9 @@ def wp(data, wt, percentiles):
     data = np.asarray(data) 
     assert len(data.shape) == 1 
     if wt is None:
-        wt = np.ones(data.shape, np.float)     
+        wt = np.ones(data.shape, float)
     else: 
-        wt = np.asarray(wt, np.float) 
+        wt = np.asarray(wt, float)
         assert wt.shape == data.shape 
         assert np.greater_equal(wt, 0.0).all(), "Not all weights are non-negative." 
     assert len(wt.shape) == 1
